drm: msm: somc_panel: limit invalid touch pinstates logging
The touch pinstates are only needed for specific panels, like incell.
It is fine to not have them defined in DT.
There is no need to warn about them each time the panel is activated/suspended.
The init code anyway warns about missing these states, so information are not lost.